+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Block library scanning until search index has been recreated/upgraded
 - Fix occasional erroneous log messages when HLS channel playback times out because all clients have left
 - Fix fallback filler playback
+- Fix stream continuity when error messages are displayed
 
 ### Added
 - Add `show_genre` and `show_tag` to search index for seasons and episodes